re-organization of preliminary icon system
